ti-bsp: Change to 6.12 as the default
We are moving to 6.12 to be our new LTS, so we are switching the default
BSP to that.  You can still access the 6.6 kernel if you need to by using
the BSP system and setting TI_PREFERRED_BSP = "ti-6_6" in your
local.conf.

Current known issues:
- Hardware graphics is disabled until we get a patch that updates the
patches to work with the 6.12 kernel.  Falling back to software
rendering.
- am62xxsip-evm and j742s2-evm are not supported in the 2025.01 u-boot
yet.  They will fail to build until the u-boot recipe is updated to
point to a commit that fixes their support.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>