ti-bsp: Disable GC320 support conditionally on BSP version
Like we do with GPU support, GC320 support is conditional on the
kernel we are building and so should be disabled here for kernels
on which it is not functional yet.

We can then remove these disabling lines as the driver is updated
for each supported kernel, instead of blanket disabling in the
distro layer as done currently.

Signed-off-by: Andrew Davis <afd@ti.com>